Feature/update 1

+++ source/businesslayer/src/main/java/com/mycompany/organizationdemo/businesslayer/managers/commands/interfaces/IDepartmentCommandManager.java
@@ -6,5 +6,5 @@ public interface IDepartmentCommandManager {
 
     DepartmentDto saveSingle(DepartmentDto item);
 
-    int deleteByKey(long key);
+    long deleteByKey(long key);
 }
